refactor: extract text preprocessing to text_utils module
